S3/FaceAlignmentService/requirements.txt
dlib==19.21.0
numpy==1.18.5
Pillow==7.2.0
opencv-python-headless
requests-toolbelt